Refactoring tests not to spam a test user
Instead, it now avoids sending actual messages 99% of the time. New
models sms.OutgoingSMS and sms.IncomingSMS store sent and received
messages, but these shield tests from checking Twilio *and* shield
Twilio from REST API spam as we test stuff.

The one test that does send a real SMS now sends it to the
TWILIO_FROM_NUMBER and uses that to test that we can receive incoming
SMSes as well. So the setting TEST_PHONE_NUMBER is now defunct. The
logic of whether or not to send an acutal message resides in the
OutgoingSMS._send_via_twilio method, which checks an invisible setting
named TEST_SENDS_ACTUAL_MESSAGES. The user will never need to set this;
getattr defaults to False except when @override_settings sets it to
True for this one test.